Side/voronets/big decimal refactoring (#15)
* added tests

* Code refactoring. Replaced Double with BigDecimal

* Revert "added tests"

This reverts commit b2746693d35a65d4dc8de4a1bfe282f55733314b.

---------

+++ src/main/java/org/nastya/dto/CurrentBalanceDTO.java
@@ -2,7 +2,9 @@ package org.nastya.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Value
 public class CurrentBalanceDTO {
-    double balance;
+    BigDecimal balance;
 }
